+++ src/languages/index.ts
@@ -1,9 +1,15 @@
 import { crystal } from "./crystal";
 import { lua } from "./lua";
 import { ruby } from "./ruby";
+import { shellscript } from "./shellscript";
 import { LanguageDefinition } from "./types";
 
-export const SUPPORTED_LANGUAGES = ["ruby", "crystal", "lua"] as const;
+export const SUPPORTED_LANGUAGES = [
+  "ruby",
+  "crystal",
+  "lua",
+  "shellscript",
+] as const;
 
 export type SupportedLanguage = (typeof SUPPORTED_LANGUAGES)[number];
 
@@ -11,6 +17,7 @@ const LANGUAGE_DEFINITIONS: Record<SupportedLanguage, LanguageDefinition> = {
   ruby,
   crystal,
   lua,
+  shellscript,
 };
 
 export function isSupportedLanguage(